+++ RELEASE_NOTES.md
@@ -1,6 +1,9 @@
 Release Notes
 =============
 
+## vNext
+* Container Apps: Fix scaler spelling
+
 ## 1.7.16
 * Container Apps: Fix a bug whereby Dapr was not correctly turned on.
 * Container Apps: Support for collections of env vars, fix ACR credentials linking.